+++ webcli/src/components/sidemenu/UserIcon.tsx
@@ -9,7 +9,7 @@ import MenuItem from '@mui/material/MenuItem';
 
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 import { logoutAsync, type AuthState } from '../../features/auth/authSlice';
-import { useAppDispatch, useAppSelector } from '../../app/hooks';
+import { useAppDispatch, useAppSelector } from '../../lib/react-redux';
 
 export default function UserIcon() {
   const selector = useAppSelector<AuthState>((state) => state.auth);
