+++ webcli/src/features/auth/components/FIDO2Register.tsx
@@ -2,7 +2,7 @@ import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 import { addFIDO2Async } from '../authSlice';
-import { useAppDispatch } from '../../../app/hooks';
+import { useAppDispatch } from '../../../lib/react-redux';
 
 export default function FIDO2Register({ onSuccess }: { onSuccess: () => void }) {
   const dispatch = useAppDispatch();
