+++ webcli/src/features/auth/components/login/MFASolutionList.tsx
@@ -6,7 +6,7 @@ import FingerprintIcon from '@mui/icons-material/Fingerprint';
 import PasswordIcon from '@mui/icons-material/Password';
 import LockClockIcon from '@mui/icons-material/LockClock';
 import { type AuthState, selectMFASolution, mfaCancelAsync } from '../../authSlice';
-import { useAppDispatch, useAppSelector } from '../../../../app/hooks';
+import { useAppDispatch, useAppSelector } from '../../../../lib/react-redux';
 import { mfasolution2name } from '../../solution';
 import { ExhaustiveError } from '../../../../utils/assert';
 import { fido2LoginAsync } from '../../thunk/fido2LoginAsync';
