+++ webcli/src/features/file/components/SearchInput.tsx
@@ -2,7 +2,7 @@ import TextField from '@mui/material/TextField';
 import InputAdornment from '@mui/material/InputAdornment';
 import SearchIcon from '@mui/icons-material/Search';
 import { changeActiveFileGroupSearch } from '../fileSlice';
-import { useAppDispatch, useAppSelector } from '../../../app/hooks';
+import { useAppDispatch, useAppSelector } from '../../../lib/react-redux';
 
 function SearchInput() {
   const dispatch = useAppDispatch();
