+++ webcli/src/features/file/components/TagEditor.tsx
@@ -1,7 +1,7 @@
 import Autocomplete, { createFilterOptions } from '@mui/material/Autocomplete';
 import TextField from '@mui/material/TextField';
 import type { FilterOptionsState } from '@mui/material/useAutocomplete';
-import { useAppSelector, useAppDispatch } from '../../../app/hooks';
+import { useAppSelector, useAppDispatch } from '../../../lib/react-redux';
 
 import type { FileState } from '../fileSlice';
 import { createDiffAsync } from '../fileSlice';
