+++ webcli/src/features/language/LangSelector.tsx
@@ -1,7 +1,7 @@
 import MenuItem from '@mui/material/MenuItem';
 import Select, { type SelectChangeEvent } from '@mui/material/Select';
 import { useTranslation } from 'react-i18next';
-import { useAppDispatch, useAppSelector } from '../../app/hooks';
+import { useAppDispatch, useAppSelector } from '../../lib/react-redux';
 import { changeLanguage } from './languageSlice';
 import { langSet } from './languageState';
