+++ webcli/src/lib/react-redux.ts
@@ -1,6 +1,6 @@
 import type { TypedUseSelectorHook } from 'react-redux';
 import { useDispatch, useSelector } from 'react-redux';
-import type { AppDispatch, RootState } from './store';
+import type { AppDispatch, RootState } from '../app/store';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
